Move ConsolePlayerIO to human/console subpackage and exclude from coverage
