+++ src/main/kotlin/werewolf/lodge/AnthropicLodge.kt
@@ -6,7 +6,7 @@ import werewolf.ai.Instruction
 import werewolf.ai.anthropic.AnthropicLanguageModel
 import werewolf.game.Player
 import werewolf.game.Role
-import werewolf.human.ConsolePlayerIO
+import werewolf.human.console.ConsolePlayerIO
 import werewolf.human.HumanPlayer
 
 object AnthropicLodge : Lodge() {
