+++ src/main/kotlin/werewolf/lodge/CpuLodge.kt
@@ -2,7 +2,7 @@ package werewolf.lodge
 
 import werewolf.game.Player
 import werewolf.game.Role
-import werewolf.human.ConsolePlayerIO
+import werewolf.human.console.ConsolePlayerIO
 import werewolf.human.HumanPlayer
 
 abstract class CpuLodge : Lodge() {
